Merge staging into main

+++ setup/setup.sh
@@ -1,17 +1,42 @@
 #!/bin/bash
 
+# Check if vectorsecrets.txt exists
+if [ ! -f vectorsecrets.txt ]; then
+    echo "vectorsecrets.txt does not exist. Please create it and try again."
+    exit 1
+fi
+
 sudo apt update
 sudo apt install wget unzip -y
-
-# Common dependencies
-bash common.sh
 
 # Vectornav Dependencies
 bash vnav.sh
 
-# Rosdep
-# sudo apt-get install python3-rosdep2 -y
-# rosdep install --from-paths ../autonav_ws/src --ignore-src -y --rosdistro=humble
+# Steam Controller Dependencies
+bash steam.sh
+
+# ImGUI Dependencies
+sudo apt install libglfw3-dev libglew-dev -y
+
+# Python deps
+sudo apt install python3-pip -y
+pip3 install python-can[serial]
+
+# Copy the udev rules to the correct location
+sudo cp autonav.rules /etc/udev/rules.d/autonav.rules
+
+# Reload udev
+sudo service udev reload
+sleep 2
+sudo service udev restart
+
+# Copy services
+sudo cp autonav.service /etc/systemd/system/autonav.service
+sudo cp autonav_service.sh /usr/bin/autonav_service.sh
+
+# chmod time :D
+sudo chmod +x /usr/bin/autonav_service.sh
+sudo chmod 644 /etc/systemd/system/autonav.service
 
 # boot time
 echo "Use 'systemctl enable autonav' to enable the service at boot time."
